Extract PageCodec to a new file
