+++ tablet/src/main/scala/wvlet/airframe/tablet/msgpack/CollectionCodec.scala
@@ -15,12 +15,10 @@ package wvlet.airframe.tablet.msgpack
 
 import java.util
 
-import org.msgpack.core.{MessagePack, MessagePacker, MessageUnpacker}
+import org.msgpack.core.{MessagePacker, MessageUnpacker}
 import org.msgpack.value.StringValue
-import wvlet.airframe.tablet.{Page, Schema}
-import wvlet.surface.Surface
 import org.msgpack.value.ValueFactory._
-import wvlet.airframe.tablet.Schema.ColumnType
+import wvlet.surface.Surface
 
 import scala.collection.JavaConverters._
 
@@ -31,64 +29,6 @@ object CollectionCodec {
 
   implicit class RichString(s: String) {
     def toValue: StringValue = newString(s)
-  }
-
-  val schemaCodec = MessageCodec.of[Schema]
-
-  object ColumnTypeCodec extends MessageCodec[ColumnType] {
-    override def pack(p: MessagePacker, v: ColumnType): Unit = {
-      p.packString(v.name)
-    }
-    override def unpack(u: MessageUnpacker, v: MessageHolder): Unit = {
-      val columnType = u.unpackString()
-      ColumnType.unapply(columnType) match {
-        case Some(c) => v.setObject(c)
-        case None    => v.setIncompatibleFormatException(s"Invalid ColumnType value: ${columnType}")
-      }
-    }
-  }
-
-  val PAGE_CODEC_ID: Byte = 0
-
-  case class PageCodec[A](surface: Surface, elementCodec: MessageCodec[A]) extends MessageCodec[Page[A]] {
-    override def pack(p: MessagePacker, v: Page[A]): Unit = {
-      // Encode page data
-      // TODO reuse buffer
-      val pagePacker = MessagePack.newDefaultBufferPacker()
-      schemaCodec.pack(pagePacker, v.schema)
-      SeqCodec(surface, elementCodec).pack(pagePacker, v.data)
-
-      // Encode page using ext type
-      val pageContent = pagePacker.toByteArray
-      p.packExtensionTypeHeader(PAGE_CODEC_ID, pageContent.length)
-      p.addPayload(pageContent)
-    }
-
-    override def unpack(u: MessageUnpacker, v: MessageHolder): Unit = {
-      val extHeader = u.unpackExtensionTypeHeader()
-      val extType   = extHeader.getType
-
-      val data = u.readPayloadAsReference(extHeader.getLength)
-      if (extType != PAGE_CODEC_ID) {
-        v.setIncompatibleFormatException(s"Unsuppoted ext type ${extType}. Expected = ${PAGE_CODEC_ID}")
-      } else {
-        val pageUnpacker = MessagePack.newDefaultUnpacker(data.sliceAsByteBuffer())
-        schemaCodec.unpack(pageUnpacker, v)
-        if (v.isNull) {
-          v.setIncompatibleFormatException(s"Failed to read schema")
-        } else {
-          val schema = v.getLastValue.asInstanceOf[Schema]
-          // Read data
-          SeqCodec(surface, elementCodec).unpack(pageUnpacker, v)
-
-          if (v.isNull) {
-            v.setIncompatibleFormatException(s"Failed to read page content")
-          } else {
-            v.setObject(Page(schema, v.getLastValue.asInstanceOf[Seq[_]]))
-          }
-        }
-      }
-    }
   }
 
   case class SeqCodec[A](surface: Surface, elementCodec: MessageCodec[A]) extends MessageCodec[Seq[A]] {
